Enhancement to application and platform status displayed.
Allow for overall Platform and individual applications to be controlled
separately via the JSON file. Each application has it overall status that
can be changed and display. If the overall Platform Status is not active
the application status are overridden to 'NOT ACTIVE' and the restored by
date and time is displayed.

The individual application status can be used at a high level to indicate
problems with specific applications to communicate to end users if that
application should be used even if it accessible on the Web UI.