Add new functionality to /nas_notify endpoint.
Platform outage form when sent updates the JSON data to set the overall
status of the NAS Platform to "NOT ACTIVE" and the restored_by to the
Outage end datetime used in the form.

This page also now has a simple toggle switch that can be used to set
the overall status of the NAS Platform back to "ACTIVE" when needed. This
allows all team members of the SA3 Core Automation Team to update the
overall Status without having to access the Openshift Namespace in order
to maintain this status during outage periods.