Wrap calling of SetConfigMaps() in app_context since the current_app is imported there in order to do logging.
Otherwise following error will be seen along with a RunTimeError: Working outside of application context exception will be seen.

"This typically means that you attempted to use functionality that needed
to interface with the current application object in some way. To solve
this, set up an application context with app.app_context().  See the
documentation for more information."